requirements.txt
discord.py 
asyncio
google_trans_new
pymongo
requests
